abstract asset storage
The license for open-web-start is BSD 3-clause. The third clause
essentially prohibits non-copyright-holders from using the project
'trademarks'. While it calls out only the name of the project/copyright
holders, for good measure I don't want to use any of the assets if I
package this in another binary.

This change abstracts the GUI assets so the packager is responsible for
provisioning them. This isn't the *best* interface, since the set of
assets is not encoded in the interface, meaning a new version could
break compatibility without the compiler catching it. But, it works well
enough. We could always extend the interface to have explicit getters in
the future if it proves problematic.